+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the desktop application will be documented in this file.
 
+## Unreleased
+
+### Features
+
+- Make correction dictionaries manageable (de92349)
+- Make settings and local polish easier to use (df5b2bc)
+
 ## v0.6.5 (2026-06-08)
 
 ### Features